GREEN Parse String to Create Purchase object
Parse very simple that ignores the penultimate token "at" and passes to
the constructor Purchase the quantity to the first token of the string,
the description to the second token of the string and the price last token of the String.
Any word between the quantity and the penultimate token "at" is part of
the description.
METHOD LIMIT: doesn't check the format of the string.