GREY Refactor
To be able to add any number of tax object without changing the previous
TaxCalculator Class, I have created a Product Interface to be able to
wrap the base product of the Purchase with any nuber of Tax Object.
In order to properly wrap the base product with the necessary taxes, I
created the FactoryTax Class.
In the future, the FactoryTax may require a refactor to be more
versatile with the addition of new Tax.
The Purchase Class for how the problem has evolved may no longer extend
the ProductBase Class (to be verified).