GREEN ImportTax Test
I created the ImportTax Class that extends TaxDecorator. I added the
boolean imported attribute to the ProductBase Class and its methods to
set it in the constructor and return it to the outside of the class.
Also I modified the FactoryTax Class to add the decoration to the
product. This Class will probably require a refactor activity.